requirements.txt
discord.py==1.2.2
requests==2.3.0
six==1.10.0